requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
#
# Use Python 3.12 or 3.13. On 3.14, pydantic-core often fails to build (PyO3).
# See docs/PYTHON_VERSION.md

Django>=5.1.4
gunicorn>=23.0.0
whitenoise>=6.8.0
psycopg[binary]>=3.2.0
django-cors-headers>=4.7.0
djangorestframework>=3.15.0
django-filter>=24.0
uvicorn>=0.34.0
logfire==3.24.0
structlog>=24.0.0
openai==1.95.1
pydantic==2.11.7
PyJWT==2.10.1
cryptography>=42.0.0
certifi>=2024.0.0
python-dotenv==1.1.1
supabase==2.16.0
letta_client>=0.1.198
google-auth==2.38.0
google-cloud-storage==3.2.0
